Preload fonts differently again, change font files
Red herring on this method being too broad - works fine; the fonts are
referenced in index.css so will be loaded on all pages that use it.

The old font files from the Ubutu site were throwing errors in the
console for some reason; these are from Google Fonts and seem better.